Add call to jl_init and the linker flags needed for this to work
We need to link against the Julia library in two places:

1. When creating the SiMoCoJulia library

2. When building the ESMX executable

For (1), we can use the results of FindJulia.cmake. This is needed for
the build of the SiMoCoJulia library to work.

(2) isn't needed for the build to work, but without it we get a runtime
error: "Invalid argument  - Unable to add component 'MDL' to driver via
shared object: ./install/lib/libSiMoCoJulia.*".

Here, we resolve (2) by adding LDFLAGS to the environment before
invoking the build; this gets picked up by the CMake build of the ESMX
driver. This would actually be enough to resolve (1) as well, but for
symmetry with how we add the include flags and for illustrative
purposes, we add an explicit target_link_libraries call in creating the
SiMoCoJulia library, even though this leads to some redundancy.